Improved cython core functions:
1. Allocate memory for DP reuse.
2. Reduced float64 to 32 for smaller memory footprint.
3. RefAligner for efficient reuse of the reference for multiple reads.
4. Can perform core alignment entirely in C. 
5. Everything is faster.

+++ nwflex/validation.py
@@ -62,9 +62,9 @@ def nwg_global(
     gs, ge = gap_open, gap_extend
 
     # three Gotoh layers
-    Yg = np.full((n + 1, m + 1), -np.inf, dtype=float)
-    M  = np.full((n + 1, m + 1), -np.inf, dtype=float)
-    Xg = np.full((n + 1, m + 1), -np.inf, dtype=float)
+    Yg = np.full((n + 1, m + 1), -np.inf, dtype=np.float32)
+    M  = np.full((n + 1, m + 1), -np.inf, dtype=np.float32)
+    Xg = np.full((n + 1, m + 1), -np.inf, dtype=np.float32)
 
     M[0, 0] = 0.0
